为管理员模式下的"确认变更"等按键添加了保护机制，当编号TextBox中的数据时空的时候按键不进行任何操作

+++ WpfApp1/Tab4.cs
@@ -42,6 +42,9 @@ namespace WpfApp1
 
         private void ShanChuDianWei_Click(object sender, RoutedEventArgs e)
         {
+            if (BianHao.Text == "")//如果TextBox中没有数据就不进行任何操作
+                return;
+
             string command_str = "delete from table3_jiedian where id=" + BianHao.Text + ";";
             MySqlHelper.GetDataSet("Database='" + ShuJuKu.ShuJuKu_Name + "';Data Source='localhost';User Id='root';Password='123456';charset='utf8';pooling=true",
                                                   CommandType.Text, command_str, null);
@@ -51,6 +54,9 @@ namespace WpfApp1
 
         private void ZengJiaBianGeng_Button_Click(object sender, RoutedEventArgs e)
         {
+            if (BianHao.Text == "")//如果TextBox中没有数据就不进行任何操作
+                return;
+
             ZengJiaBianGeng(Convert.ToInt16(BianHao.Text), JianCeQiTi.Text, AnZhuangWeiZhi.Text, AnZhuangShiJina.Text, GaoXianBaoJing.Text, DiXianBaoJing.Text);
 
             //根据当前新增节点的坐标更新数据库中的数据
@@ -98,6 +104,8 @@ namespace WpfApp1
 
         private void QueRenBianGeng_Button_Click(object sender, RoutedEventArgs e)
         {
+            if (BianHao.Text == "")//如果TextBox中没有数据就不进行任何操作
+                return;
             Update_Information_Jiedian(Ellipse_Array_tab4, Convert.ToInt16(BianHao.Text), JianCeQiTi.Text, AnZhuangWeiZhi.Text, AnZhuangShiJina.Text, GaoXianBaoJing.Text, DiXianBaoJing.Text);
             Init_Jiedian_DisplayOrNot();//刷新所有节点
         }
